Struct phantom_newtype::AmountOf

pub struct AmountOf<Unit, Repr>(_, _);

AmountOf<Unit> provides a type-safe way to keep an amount of some Unit.

E.g. the following code must not compile:

use phantom_newtype::AmountOf;

// These structs are just tags and have no semantic meaning.
enum Apples {}
enum Oranges {}

let trois_pommes = AmountOf::<Apples, u64>::from(3);
let three_oranges = AmountOf::<Oranges, u64>::from(3);

assert_eq!(trois_pommes, three_oranges)

AmountOf<Unit, Repr> defines common boilerplate to make type-safe amounts more convenient. For example, you can compare amounts:

use phantom_newtype::AmountOf;

enum MetricApple {}
type Apples = AmountOf<MetricApple, u64>;

assert!(Apples::from(3) < Apples::from(5));
assert_eq!(false, Apples::from(3) > Apples::from(5));
assert!(Apples::from(3) != Apples::from(5));
assert!(Apples::from(5) == Apples::from(5));
assert_eq!(false, Apples::from(5) != Apples::from(5));

assert_eq!(vec![Apples::from(3), Apples::from(5)].iter().max().unwrap(),
           &Apples::from(5));

You can do simple arithmetics with amounts:

use phantom_newtype::AmountOf;

enum MetricApple {}
type Apples = AmountOf<MetricApple, u64>;

let mut x = Apples::from(5);
let y = Apples::from(3);

assert_eq!(x + y, Apples::from(8));
assert_eq!(x - y, Apples::from(2));
assert_eq!(x.increment(), Apples::from(6));
assert_eq!(x.decrement(), Apples::from(4));

// Variables are unmodified.
assert_eq!(x, Apples::from(5));
assert_eq!(y, Apples::from(3));

// But add assign and subtract assign will mutate the variable.
x += y;
assert_eq!(x, Apples::from(8));
x -= y;
assert_eq!(x, Apples::from(5));
x.inc_assign();
assert_eq!(x, Apples::from(6));
x.dec_assign();
assert_eq!(x, Apples::from(5));

assert_eq!(Apples::from(55), (1..=10_u64).map(Apples::from).sum());

Multiplication of amounts is not supported: multiplying meters by meters gives square meters. However, you can scale an amount by a scalar; divide amounts; or divide amounts by scalars:

use phantom_newtype::AmountOf;

enum Meter {}

type Meters = AmountOf<Meter, u64>;

let mut x = Meters::from(5);

assert_eq!(x * 3, Meters::from(15));
assert_eq!(1, x / x);
assert_eq!(3, (x * 3) / x);
// Note: integer division, because the representation is u64.
assert_eq!(Meters::from(1), x / 3);

// Variable is unmodified.
assert_eq!(x, Meters::from(5));

// But multiply assign and div assign will mutate it.
x *= 3;
assert_eq!(x, Meters::from(15));
x /= 5;
assert_eq!(x, Meters::from(3));

Note that the unit is only available at compile time, thus using AmountOf instead of u64 doesn't incur any runtime penalty:

use phantom_newtype::AmountOf;

enum Meter {}

type Meters = AmountOf<Meter, u64>;

let ms = Meters::from(10);
assert_eq!(std::mem::size_of_val(&ms), std::mem::size_of::<u64>());

Amounts can be serialized and deserialized with serde. Serialized forms of Amount<Unit, Repr> and Repr are identical.

use phantom_newtype::AmountOf;
use serde::{Serialize, Deserialize};
use serde_json;

enum Meter {}
type Meters = AmountOf<Meter, u64>;

let repr: u64 = 10;
let m_10 = Meters::from(repr);
assert_eq!(serde_json::to_string(&m_10).unwrap(), serde_json::to_string(&repr).unwrap());

let copy: Meters = serde_json::from_str(&serde_json::to_string(&m_10).unwrap()).unwrap();
assert_eq!(copy, m_10);

You can also declare constants of AmountOf<Unit, Repr> using new function:

use phantom_newtype::AmountOf;
enum Meter {}
type Distance = AmountOf<Meter, u64>;
const ASTRONOMICAL_UNIT: Distance = Distance::new(149_597_870_700);

assert!(ASTRONOMICAL_UNIT > Distance::from(0));

Implementations

impl<Unit, Repr: Copy> AmountOf<Unit, Repr>

pub fn get(&self) -> Repr

Returns the wrapped value.

use phantom_newtype::AmountOf;

enum Apple {}

let three_apples = AmountOf::<Apple, u64>::from(3);
assert_eq!(9, (three_apples * 3).get());

impl<Unit, Repr> AmountOf<Unit, Repr>

pub const fn new(repr: Repr) -> AmountOf<Unit, Repr>

new is a synonym for from that can be evaluated in compile time. The main use-case of this functions is defining constants.

impl<Unit: Default, Repr> AmountOf<Unit, Repr>

pub fn unit(&self) -> Unit

Provides a useful shortcut to access units of an amount if they implement the Default trait:

use phantom_newtype::AmountOf;

#[derive(Debug, Default)]
struct Seconds {}
let duration = AmountOf::<Seconds, u64>::from(5);

assert_eq!("5 Seconds", format!("{} {:?}", duration, duration.unit()));

impl<Unit, Repr> AmountOf<Unit, Repr> where
Unit: DisplayerOf<AmountOf<Unit, Repr>>,

pub fn display(&self) -> DisplayProxy<Self, Unit>

display provides a machanism to implement a custom display for phantom types.

use phantom_newtype::{AmountOf, DisplayerOf};
use std::fmt;

enum Cent {}
type Money = AmountOf<Cent, u64>;

impl DisplayerOf<Money> for Cent {
  fn display(amount: &Money, f: &mut fmt::Formatter<'_>) -> fmt::Result {
    write!(f, "${}.{:02}", amount.get() / 100, amount.get() % 100)
  }
}

assert_eq!(format!("{}", Money::from(1005).display()), "$10.05");

impl<Unit, Repr> AmountOf<Unit, Repr> where
Repr: Add<Output = Repr> + From<u8>,

pub fn increment(self) -> AmountOf<Unit, Repr>

Returns the amount incremented by 1.

impl<Unit, Repr> AmountOf<Unit, Repr> where
Repr: AddAssign + From<u8>,

pub fn inc_assign(&mut self)

Increments the amount by 1.

impl<Unit, Repr> AmountOf<Unit, Repr> where
Repr: Sub<Output = Repr> + From<u8>,

pub fn decrement(self) -> AmountOf<Unit, Repr>

Returns the amount decremented by 1. Like regular subtraction, panics in debug mode if Repr is an unsigned integer type and the amount is zero.

impl<Unit, Repr> AmountOf<Unit, Repr> where
Repr: SubAssign + From<u8>,

pub fn dec_assign(&mut self)

Decrements the amount by 1. Like regular subtraction, panics in debug mode if Repr is an unsigned integer type and the amount is zero.